requirements.txt
# Editable install with no version control (ai-assistant==0.1.0)
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.121.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.71.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json_repair==0.52.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-openai==1.0.2
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langfuse==2.60.10
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.2
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.3.4
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rpds-py==0.28.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.51.0
tenacity==9.1.2
tiktoken==0.12.0
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
